// web/src/schema/site.schema.d.ts
export interface GitHubConnection {
  /**
   * URL of a GitHub instance, such as https://github.com or https://github-enterprise.example.com.
   */
  url: string;
  /**
   * The type of Git URLs to use for cloning and fetching Git repositories on this GitHub instance.
   *
   * If "http", Sourcegraph will access GitLab repositories using Git URLs of the form http(s)://github.com/myteam/myproject.git (using https: if the GitHub instance uses HTTPS).
   *
   * If "ssh", Sourcegraph will access GitHub repositories using Git URLs of the form git@github.com:myteam/myproject.git. See the documentation for how to provide SSH private keys and known_hosts: https://docs.sourcegraph.com/admin/repo/add_from_git_repository#repositories-that-need-http-s-or-ssh-authentication.
   */
  gitURLType?: "http" | "ssh";
  /**
   * A GitHub personal access token with repo and org scope.
   */
  token: string;
  /**
   * TLS certificate of a GitHub Enterprise instance.
   */
  certificate?: string;
  /**
   * An array of repository "owner/name" strings specifying which GitHub or GitHub Enterprise repositories to mirror on Sourcegraph.
   */
  repos?: string[];
  /**
   * An array of strings specifying which GitHub or GitHub Enterprise repositories to mirror on Sourcegraph. The valid values are:
   *
   * - `public` mirrors all public repositories for GitHub Enterprise and is the equivalent of `none` for GitHub
   *
   * - `affiliated` mirrors all repositories affiliated with the configured token's user:
   * 	- Private repositories with read access
   * 	- Public repositories owned by the user or their orgs
   * 	- Public repositories with write access
   *
   * - `none` mirrors no repositories (except those specified in the `repos` configuration property or added manually)
   *
   * - All other values are executed as a GitHub advanced repository search as described at https://github.com/search/advanced. Example: to sync all repositories from the "sourcegraph" organization including forks the query would be "org:sourcegraph fork:true".
   *
   * If multiple values are provided, their results are unioned.
   *
   * If you need to narrow the set of mirrored repositories further (and don't want to enumerate it with a list or query set as above), create a new bot/machine user on GitHub or GitHub Enterprise that is only affiliated with the desired repositories.
   */
  repositoryQuery?: string[];
  /**
   * The pattern used to generate the corresponding Sourcegraph repository name for a GitHub or GitHub Enterprise repository. In the pattern, the variable "{host}" is replaced with the GitHub host (such as github.example.com), and "{nameWithOwner}" is replaced with the GitHub repository's "owner/path" (such as "myorg/myrepo").
   *
   * For example, if your GitHub Enterprise URL is https://github.example.com and your Sourcegraph URL is https://src.example.com, then a repositoryPathPattern of "{host}/{nameWithOwner}" would mean that a GitHub repository at https://github.example.com/myorg/myrepo is available on Sourcegraph at https://src.example.com/github.example.com/myorg/myrepo.
   *
   * It is important that the Sourcegraph repository name generated with this pattern be unique to this code host. If different code hosts generate repository names that collide, Sourcegraph's behavior is undefined.
   */
  repositoryPathPattern?: string;
  /**
   * Defines whether repositories from this GitHub instance should be enabled and cloned when they are first seen by Sourcegraph. If false, the site admin must explicitly enable GitHub repositories (in the site admin area) to clone them and make them searchable on Sourcegraph. If true, they will be enabled and cloned immediately (subject to rate limiting by GitHub); site admins can still disable them explicitly, and they'll remain disabled.
   */
  initialRepositoryEnablement?: boolean;
}
export interface GitLabConnection {
  /**
   * URL of a GitLab instance, such as https://gitlab.example.com or (for GitLab.com) https://gitlab.com.
   */
  url: string;
  /**
   * A GitLab access token with "api" and "sudo" scopes. If this token does not have "sudo" scope, then you must set `permissions.ignore` to true.
   */
  token: string;
  /**
   * The type of Git URLs to use for cloning and fetching Git repositories on this GitLab instance.
   *
   * If "http", Sourcegraph will access GitLab repositories using Git URLs of the form http(s)://gitlab.example.com/myteam/myproject.git (using https: if the GitLab instance uses HTTPS).
   *
   * If "ssh", Sourcegraph will access GitLab repositories using Git URLs of the form git@example.gitlab.com:myteam/myproject.git. See the documentation for how to provide SSH private keys and known_hosts: https://docs.sourcegraph.com/admin/repo/add_from_git_repository#repositories-that-need-http-s-or-ssh-authentication.
   */
  gitURLType?: "http" | "ssh";
  /**
   * TLS certificate of a GitLab instance.
   */
  certificate?: string;
  /**
   * An array of strings specifying which GitLab projects to mirror on Sourcegraph. Each string is a URL query string for the GitLab projects API, such as "?membership=true&search=foo".
   *
   * The query string is passed directly to GitLab to retrieve the list of projects. The special string "none" can be used as the only element to disable this feature. Projects matched by multiple query strings are only imported once. See https://docs.gitlab.com/ee/api/projects.html#list-all-projects for available query string options.
   */
  projectQuery?: string[];
  /**
   * The pattern used to generate a the corresponding Sourcegraph repository name for a GitLab project. In the pattern, the variable "{host}" is replaced with the GitLab URL's host (such as gitlab.example.com), and "{pathWithNamespace}" is replaced with the GitLab project's "namespace/path" (such as "myteam/myproject").
   *
   * For example, if your GitLab is https://gitlab.example.com and your Sourcegraph is https://src.example.com, then a repositoryPathPattern of "{host}/{pathWithNamespace}" would mean that a GitLab project at https://gitlab.example.com/myteam/myproject is available on Sourcegraph at https://src.example.com/gitlab.example.com/myteam/myproject.
   *
   * It is important that the Sourcegraph repository name generated with this pattern be unique to this code host. If different code hosts generate repository names that collide, Sourcegraph's behavior is undefined.
   */
  repositoryPathPattern?: string;
  /**
   * Defines whether repositories from this GitLab instance should be enabled and cloned when they are first seen by Sourcegraph. If false, the site admin must explicitly enable GitLab repositories (in the site admin area) to clone them and make them searchable on Sourcegraph. If true, they will be enabled and cloned immediately (subject to rate limiting by GitLab); site admins can still disable them explicitly, and they'll remain disabled.
   */
  initialRepositoryEnablement?: boolean;
  /**
   * If non-null, enables GitLab permission checks. This requires that the value of `token` be an access token with "sudo" and "api" scopes.
   */
  authorization?: {
    /**
     * Identifies the authentication provider to use to identify users to GitLab.
     */
    authnProvider: {
      /**
       * The value of the `configID` field of the targeted authentication provider.
       */
      configID: string;
      /**
       * The `type` field of the targeted authentication provider.
       */
      type: string;
      /**
       * The provider name that identifies the authentication provider to GitLab. This is the name passed to the `?provider=` query parameter in calls to the GitLab Users API.
       */
      gitlabProvider: string;
    };
    /**
     * The TTL of how long to cache permissions data. This is 3 hours by default.
     *
     * Decreasing the TTL will increase the load on the code host API. If you have X repos on your instance, it will take ~X/100 API requests to fetch the complete list for 1 user.  If you have Y users, you will incur X*Y/100 API requests per cache refresh period.
     *
     * If set to zero, Sourcegraph will sync a user's entire accessible repository list on every request (NOT recommended).
     */
    ttl?: string;
  };
}
// ...
export interface Repository {
  /**
   * Type of the version control system for this repository, such as "git"
   */
  type?: "git";
  /**
   * Clone URL for the repository, such as git@example.com:my/repo.git
   */
  url: string;
  /**
   * Display path on Sourcegraph for the repository, such as my/repo
   */
  path: string;
  links?: {
    /**
     * URL specifying where to view the repository at an external location e.g. "https://example.com/myrepo"
     */
    repository?: string;
    /**
     * URL template for specifying how to link to commits at an external location. Use "{commit}" as a placeholder for a given commit ID e.g. "https://example.com/myrepo/view-commit/{commit}"
     */
    commit?: string;
    /**
     * URL template for specifying how to link to paths at an external location. Use "{path}" as a placeholder for a given path and "{rev}" as a placeholder for a given revision e.g. "https://example.com/myrepo@{rev}/browse/{path}"
     */
    tree?: string;
    /**
     * URL template for specifying how to link to files at an external location. Use "{path}" as a placeholder for a given path and "{rev}" as a placeholder for a given revision e.g. "https://example.com/myrepo@{rev}/browse/{path}"
     */
    blob?: string;
  };
}